fix: move model_fuser to lib/shared/utils, matching internal layout
Internal's fuser lives at shared/utils/model_fuser -- a sibling of
sdk (which holds model_manager), not nested under it -- since it's a
cross-cutting utility, not a model_manager concern. Move OSS's fuser
out from under lib/model_manager/utils/torch/ to a new lib/shared/utils/
package to match, and rename schema_fuse.py to fuse_schema.py to match
internal's naming exactly (helps a future re-import from OSS back into
internal go directly file-for-file).
