+++ python/michelangelo/lib/shared/__init__.py
@@ -1,0 +1,5 @@
+"""Cross-cutting utilities shared across ``lib`` subpackages.
+
+Code here is not specific to any one ``lib`` subpackage (e.g. ``model_manager``,
+``trainer``, ``artifact_manager``) and may be depended on by several of them.
+"""
