+++ python/michelangelo/lib/shared/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Shared utility subpackages (e.g. ``model_fuser``)."""
