+++ python/michelangelo/lib/shared/utils/model_fuser/__init__.py
@@ -14,8 +14,8 @@ from .fuse import (
     fuse_models_to_torchscript,
     get_predictor_output_field_order,
 )
+from .fuse_schema import fuse_input_schema, fuse_model_schema
 from .fused_model import FusedModel
-from .schema_fuse import fuse_input_schema, fuse_model_schema
 
 __all__ = [
     "FusedModel",
