+++ python/michelangelo/lib/shared/utils/model_fuser/fused_model.py
