+++ python/michelangelo/lib/shared/utils/model_fuser/tests/fuse_schema_test.py
@@ -1,11 +1,11 @@
-"""Unit tests for ``model_manager.utils.torch.model_fuser.schema_fuse``."""
+"""Unit tests for ``lib.shared.utils.model_fuser.fuse_schema``."""
 
 from __future__ import annotations
 
 import unittest
 
 from michelangelo.lib.model_manager.schema import DataType, ModelSchema, ModelSchemaItem
-from michelangelo.lib.model_manager.utils.torch.model_fuser.schema_fuse import (
+from michelangelo.lib.shared.utils.model_fuser.fuse_schema import (
     fuse_input_schema,
     fuse_model_schema,
 )
