+++ python/michelangelo/workflow/tasks/tabular_assembler/custom/assembler.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING
 from michelangelo.lib.model_manager.constants import StorageType
 from michelangelo.lib.model_manager.packager.custom_triton import CustomTritonPackager
 from michelangelo.lib.model_manager.schema.sample_data_fuse import fuse_sample_data
-from michelangelo.lib.model_manager.utils.torch.model_fuser import fuse_model_schema
+from michelangelo.lib.shared.utils.model_fuser import fuse_model_schema
 from michelangelo.workflow.tasks.tabular_assembler._private.model_class.resolve import (
     resolve_model_class,
 )
